+++ R/waitForFiles.R
@@ -1,7 +1,7 @@
 # use list.files() here as this seems to trick the nfs cache
 # see https://github.com/mllg/batchtools/issues/85
-waitForFiles = function(path, fns, timeout = NA_real_) {
-  if (is.na(timeout))
+waitForFiles = function(path, fns, timeout = 0) {
+  if (timeout == 0)
     return(TRUE)
 
   fns = fns[!fs::file_exists(fns)]
@@ -24,25 +24,22 @@ waitForFiles = function(path, fns, timeout = NA_real_) {
   }
 }
 
-waitForFile = function(fn, timeout = NA_real_) {
-  if (is.na(timeout) || fs::file_exists(fn))
+waitForFile = function(fn, timeout = 0, must.work = TRUE) {
+  if (timeout == 0 || fs::file_exists(fn))
     return(TRUE)
 
   "!DEBUG [waitForFile]: `fn` not found via 'file.exists()'"
   timeout = timeout + Sys.time()
+  path = fs::path_dir(fn)
   repeat {
     Sys.sleep(0.5)
-    if (fn %chin% list.files(fs::path_dir(fn), all.files = TRUE))
+    if (fn %chin% list.files(path, all.files = TRUE))
       return(TRUE)
-    if (Sys.time() > timeout)
-      stopf("Timeout while waiting for file '%s'", fn)
+    if (Sys.time() > timeout) {
+      if (must.work)
+        stopf("Timeout while waiting for file '%s'", fn)
+      return(FALSE)
+    }
   }
 }
 
-waitForResults = function(reg, ids) {
-  waitForFiles(
-    fs::path(reg$file.dir, "results"),
-    sprintf("%i.rds", .findDone(reg, ids)$job.id),
-    reg$cluster.functions$fs.latency
-  )
-}