+++ NEWS.md
@@ -1,6 +1,7 @@
 # batchtools 0.9.11
 
 * Removed deprecated function `chunkIds()`.
+* New default for argument `fs.timeout` in the cluster function constructor is `0` (was `NA` before).
 
 # batchtools 0.9.10